refactor: division of code into modules
