+++ index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Virtual keyboard</title>
     <link rel="stylesheet" href="./style.css" />
-    <script type="module" src="script.js" defer></script>
+    <script type="module" src="index.js" defer></script>
   </head>
   <body></body>
 </html>
