+++ modules/helper.js
@@ -1,0 +1,6 @@
+export const addAndRemoveActive = (code) => {
+  document.querySelector(code).classList.add('active');
+  setTimeout(() => {
+  document.querySelector(code).classList.remove('active');
+  }, 200);
+};
